# api/openapi.yaml
openapi: 3.1.0
info:
  title: agentibrain-kernel HTTP API
  version: 0.1.0
  description: |
    Canonical HTTP contract for the agentibrain kernel. Consumed by:

    - `agentihooks` hooks (brain_adapter → /feed, brain_writer_hook → /marker,
      amygdala_hook → /signal).
    - KB MCP tools (`kb_search`, `kb_brief`, `kb_dispatch`).
    - Humans + other agents for ingest and retrieval.

    **Auth:** bearer token via `Authorization: Bearer $KB_ROUTER_TOKEN`. The
    token is issued at `agentibrain init` time and stored in `~/.agentibrain/.env`.

    **Status:** v0.1.0 ships `/health`, `/ingest`, `/search`, `/brief`,
    `/dispatch`. The fleet-facing endpoints (`/feed`, `/signal`, `/marker`,
    `/tick`) land in Phase 7 when the agentihooks connectors are rewritten.
servers:
  - url: http://localhost:8102
    description: Default local brain-api port (agentibrain init default).
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    IngestRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: Free-form message; router classifies and fans out.
        title:
          type: string
        tags:
          type: array
          items: { type: string }
    IngestResponse:
      type: object
      properties:
        obsidian_ref:
          type: string
          nullable: true
        ingest_batch:
          type: string
        artifacts:
          type: array
          items: { type: object }
...
    MarkerRequest:
      type: object
      required: [type, content]
      properties:
        type:
          type: string
          enum: [lesson, milestone, signal, decision]
        content: { type: string }
        severity:
          type: string
          enum: [nuclear, critical, warning, info, resolved]
        source: { type: string }
security:
  - bearerAuth: []
paths:
...
  /ingest:
    post:
      summary: Classify + fan-out an ingest payload.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IngestRequest" }
      responses:
        "200":
          description: Ingest summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IngestResponse" }
...
  /marker:
    post:
      summary: Record a lesson / milestone / signal / decision marker (Phase 7).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MarkerRequest" }
      responses:
        "202":
          description: Marker accepted.
...